Move AnimationQueue from base to core
